move tests with more RAM requirement to a separate file
